[LCR-199] refactor: 공통 Enum Common 모듈로 분리
