+++ popi-auth-service/src/main/java/com/lgcns/dto/AccessTokenDto.java
@@ -1,6 +1,6 @@
 package com.lgcns.dto;
 
-import com.lgcns.domain.MemberRole;
+import com.lgcns.enums.MemberRole;
 
 public record AccessTokenDto(Long memberId, MemberRole role, String accessTokenValue) {
     public static AccessTokenDto of(Long memberId, MemberRole role, String accessTokenValue) {
