+++ popi-auth-service/src/main/java/com/lgcns/util/JwtUtil.java
@@ -2,10 +2,10 @@ package com.lgcns.util;
 
 import static com.lgcns.constants.SecurityConstants.TOKEN_ROLE_NAME;
 
-import com.lgcns.domain.MemberRole;
 import com.lgcns.dto.AccessTokenDto;
 import com.lgcns.dto.RefreshTokenDto;
 import com.lgcns.dto.RegisterTokenDto;
+import com.lgcns.enums.MemberRole;
 import com.lgcns.infra.jwt.JwtProperties;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
