--- popi-member-service/src/main/java/com/lgcns/domain/MemberRole.java
@@ -1,14 +1,0 @@
-package com.lgcns.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum MemberRole {
-    ADMIN("ROLE_ADMIN"),
-    USER("ROLE_USER"),
-    ;
-
-    private final String role;
-}
